refactor: edit Valeu type to int64 instead float

+++ models/bill.go
@@ -20,9 +20,9 @@ type Bills struct {
 }
 
 type BillInput struct {
-	Value    float64 `json:"value" binding:"required,gt=0"`
-	Category string  `json:"category" binding:"required"`
-	Status   string  `json:"status" binding:"required,oneof='paid' 'unpaid' 'pending'"`
+	Value    int64  `json:"value" binding:"required,gt=0"`
+	Category string `json:"category" binding:"required"`
+	Status   string `json:"status" binding:"required,oneof='paid' 'unpaid' 'pending'"`
 }
 
 func (b *Bills) BeforeCreate(d *gorm.DB) (err error) {
